+++ CHANGES.md
@@ -25,6 +25,7 @@ Changes in 12.5
 - JS required group of checkboxes in Inputs.php & MultipleCheckboxInput.js
 - JS remove inline code from PasswordInput() in Inputs.php & PasswordInput.js
 - JS remove inline code from DeletePrompt() & Prompt() + add .button-prompt-cancel CSS class in Prompts.php & Prompt.js
+- JS remove inline code from Search() in Search.fnc.php & Search.js
 
 Changes in 12.4.3
 -----------------